% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{blue}
\alias{blue}
\title{Keep blue in the picture}
\usage{
blue(width, height, filter_bitmap)
}
\arguments{
\item{width}{the width of the picture}

\item{height}{the height of the picture}

\item{filter_bitmap}{the raw() vector from the original picture (gerenated by readImage)}
}
\value{
Blue in raw form for a specific pixel
}
\description{
Keep blue in the picture
}
