% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{grey}
\alias{grey}
\title{Calculate the greyscale for the chosen pixel}
\usage{
grey(width, height, filter_bitmap)
}
\arguments{
\item{width}{the width of the picture}

\item{height}{the height of the picture}

\item{filter_bitmap}{the raw() vector from the original picture (gerenated by readImage)}
}
\value{
the color in raw form for a specific pixel
}
\description{
Calculate the greyscale for the chosen pixel
}
