% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImage.R
\name{readImage}
\alias{readImage}
\title{Read the image of interest}
\usage{
readImage(image_path = "figures/Testimage1.jpg")
}
\arguments{
\item{image_path}{the pathway of the image you want to manipulate}
}
\value{
image display on the R window
}
\description{
Read the image of interest
}
