% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{red}
\alias{red}
\title{Keep red in the picture}
\usage{
red(width, height, filter_bitmap)
}
\arguments{
\item{width}{the width of the picture}

\item{height}{the height of the picture}

\item{filter_bitmap}{the raw() vector from the original picture (gerenated by readImage)}
}
\value{
red in raw form for a specific pixel
}
\description{
Keep red in the picture
}
