% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelchange.R
\name{pixelchange}
\alias{pixelchange}
\title{loop function to change the color pixel by pixel}
\usage{
pixelchange(
  ori_bitmap = image_bitmap,
  colorscale = "grey",
  output = "FALSE",
  outpathway = "figures/output.jpg"
)
}
\arguments{
\item{ori_bitmap}{the raw() vector generated by readImage}

\item{colorscale}{the filter you want to apply}

\item{output}{whether you want up output the new image}

\item{outpathway}{the pathway for output image default save to "wd/figures/output.jpg"}
}
\value{
raw() vector of the new image with the filter chosen
}
\description{
default setting grey filter
create a new object for the return e.g. new_obj <- pixelchange()
use magick::fun to export the new_obj
}
\examples{

# import_image_bitmap <- readImage("figures/Testimage1.jpg")
# pixelchange( ori_bitmap=import_image_bitmap, "grey", "FALSE",)

}
